requirements.txt
Flask==2.2.5
Flask-RESTX==1.2.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
PyMySQL
